Move to using ClusterData for Vagrantfile (and split into three BACH::ClusterData modules)
