+++ repxe_host.rb
@@ -55,6 +55,7 @@ require 'timeout'
 require 'optparse'
 
 require_relative 'lib/cluster_data'
+require_relative 'lib/chef_node'
 include BACH::ClusterData
 include BACH::ClusterData::ChefNode
